app/requirements.txt
# Cython==0.28.5
Django>=2.0,<3.0
psycopg2>=2.7,<3.0

# These should be okay to update minors
numpy>=1.14,<1.14.99
pandas>=0.23,<0.23.99

pinax-eventlog>=2.0.3,<2.99
django-crispy-forms
xlrd>=1.2.0,<1.3
django-xml
requests
pinax-eventlog>=2.0.3,<2.99

mysqlclient>=1.3.13,<1.3.99

biopython
moca